requirements.txt
# OMEGA v4 TITANIUM PRO - Python Dependencies

# Core Framework
streamlit>=1.30.0

# Data Processing
pandas>=2.0.0

# AI & Machine Learning
google-generativeai>=0.3.0

# AI Router Dependencies
requests>=2.31.0
duckduckgo-search>=6.0.0

# Document Processing (for AI Router)
PyPDF2>=3.0.0
python-docx>=1.1.0
openpyxl>=3.1.0
mutagen>=1.47.0

# File Handling
rarfile>=4.0
py7zr>=0.20.0

# Environment Variables
python-dotenv>=1.0.0

# Image Processing (optional, for thumbnails)
Pillow>=10.0.0
